% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/county_report.R
\name{county_report}
\alias{county_report}
\title{County Report}
\usage{
county_report(counties, region_name, from, to, output)
}
\arguments{
\item{counties}{list of county codes from CDW}

\item{region_name}{name you'd like to give the regional area as it appears in the report}

\item{from}{start date in yyyymmdd format}

\item{to}{end date in yyyymmdd format}

\item{output}{file location and name for report}
}
\description{
This function creates a Regional Report for specified counties and outputs the report as an html file.
}
\examples{
# This would output the county report for Southern California from January 1, 2014 to June 30, 2018
county_report(counties = c("CA037", "CA059", "CA065", "CA071", "CA073", "CA083", "CA111"), region_name = "Southern California", from = 20140101, to = 20180630, output = "southern_ca_report")
}
