% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regional_report.R
\name{regional_report}
\alias{regional_report}
\title{Regional Report}
\usage{
regional_report(msa, msa_name, from, to, output)
}
\arguments{
\item{msa}{MSA code from CDW}

\item{msa_name}{name you'd like to give MSA as it appears in the report}

\item{from}{start date in yyyymmdd format}

\item{to}{end date in yyyymmdd format}

\item{output}{file location and name for report}
}
\description{
This function creates a Regional Report for a specified MSA and outputs the report as an html file.
}
\examples{
# This would output the regional report for Boston from January 1, 2014 to June 30, 2018
regional_report(msa = 14460, msa_name = "Boston", from = 20140101, to = 20180630, output = "my_file/boston_report")
}
