% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regional_school_report.R
\name{regional_school_report}
\alias{regional_school_report}
\title{Regional School Report}
\usage{
regional_school_report(msa, msa_name, school_code, school_name, from, to,
  output)
}
\arguments{
\item{msa}{MSA code from CDW}

\item{msa_name}{name you'd like to give MSA as it appears in the report}

\item{school_code}{school code from CDW}

\item{school_name}{school name as you'd like it to appear in the report}

\item{from}{start date in yyyymmdd format}

\item{to}{end date in yyyymmdd format}

\item{output}{file location and name for report}
}
\description{
This function creates a Regional School Report for specified MSA and graduating school and outputs the report as an html file.
}
\examples{
# This would output the report for Boston and the College of Engineering from January 1, 2014 to June 30, 2018
school_report(msa = 14460, msa_name = "Boston", school_code = "EN", school_name = "College of Engineering", from = 20140101, to = 20180630, output = "my_file/boston_engineering_report")
}
