% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/school_report.R
\name{school_report}
\alias{school_report}
\title{School Report}
\usage{
school_report(school_code, school_name, from, to, output)
}
\arguments{
\item{school_code}{school code from CDW}

\item{school_name}{school name as you'd like it to appear in the report}

\item{from}{start date in yyyymmdd format}

\item{to}{end date in yyyymmdd format}

\item{output}{file location and name for report}
}
\description{
This function creates a School Report for a specified graduating school and outputs the report as an html file.
}
\examples{
# This would output the school report for the College of Engineering from January 1, 2014 to June 30, 2018
school_report(school = EN, school_name = "College of Engineering", from = 20140101, to = 20180630, output = "my_file/engineering_report")
}
